+++ ft_printf/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kryxaurus <kryxaurus@student.42.fr>        +#+  +:+       +#+        */
+/*   By: khlavaty <khlavaty@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/20 18:11:09 by khlavaty          #+#    #+#             */
-/*   Updated: 2023/03/01 01:00:58 by kryxaurus        ###   ########.fr       */
+/*   Updated: 2023/03/01 16:07:44 by khlavaty         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,10 +24,10 @@ int		fts_check_formt(va_list args, char formt);
 int		ft_putstring(char *str);
 int		ft_putnbr(int n);
 int		ft_putnbrunsig(unsigned int n);
-int		ft_puthex(long n, char formt);
+int		ft_puthex(unsigned long long n, char formt);
 size_t	ft_strlen(const char *str);
-int		ft_putbase(long n, char *base, int base_len);
-int		ft_putnbrbase(long n, char *base);
+int		ft_putbase(unsigned long long n, char *base, int base_len);
+int		ft_putnbrbase(unsigned long long n, char *base);
 int		ft_mem(char *mem_c, uintptr_t mem);
 int		ft_putmemory(uintptr_t mem);
